refactor(core): drop usecase from adapter deps, migrate interface_adapters to adapters
ElectrumServersAdapter (electrum), ElectrumServerAdapter (blockchain,
wallet) previously depended on GetElectrumServersToUseUsecase — an
application-layer use case invoked from secondary adapters, reversing
the layer-direction rule. The usecase served these adapters only.

Absorb the orchestration (fetch + filter custom + sort by priority)
into ElectrumServerRepository.fetchActiveServers as a query method.
The two non-electrum adapters compose it with ElectrumSettingsRepository
and SettingsRepository (for tor proxy) inline. Delete the usecase plus
its request/response DTOs.

Relocate all touched files from the legacy interface_adapters/ folder
to the canonical adapters/ folder (electrum, wallet — matching the
blockchain move already in flight). Remaining interface_adapters/
usage in tor/mempool/swaps/status is out of scope for this PR.
