+++ lib/core/electrum/adapters/drift_electrum_settings_repository.dart
