+++ lib/core/electrum/adapters/environment_adapter.dart
