--- lib/core/electrum/application/dtos/requests/get_electrum_servers_to_use_request.dart
@@ -1,7 +1,0 @@
-import 'package:bb_mobile/core/electrum/domain/value_objects/electrum_server_network.dart';
-
-class GetElectrumServersToUseRequest {
-  final ElectrumServerNetwork network;
-
-  GetElectrumServersToUseRequest({required this.network});
-}
